Working of connection
*Moving incoming connetions of the peer from pointers collection
of pointers to collection of objects.
GOOD: Will be faster due continue mem
BAD: Well this may create a huge overhead when we will start to clearing
a lost connection somewhere in middle, to many copys of entier vector. But
should't be that offted Will check when start to massuer stuff
TDO IMPLACE

Signed-off-by: Daniel Mor <onecoolterran@gmail.com>